Move Hive metastore creation to TokenDelegationThriftMetastoreFactory
This factory encapsulates all logic for thrift hive metastore client
creation.
